Add support for a user-defined transport CA certificate and key (#4053)
Structure of the CA secret is analogous to the HTTP certificate secret we already support and it expects a tls.crt and a tls.key entry to contain the certificate as well as the private key in PEM format.

If specified the user provided CA certificate will be used instead of a self-signed certificate authority generated by the operator.

When a custom CA is specified the operator will remove the self-signed CA secret and not reconcile it anymore. The custom certificate will be parsed and validated in the reconciliation loop and not at admission. Any parsing errors will be surfaced as events (in addition to appearing in the logs)