E2E test verify peer CA certificate during TLS handshake
Work around the issue that peer certificates are only populated in connection state after a successful handshake and that Elasticsearch 6.x is more restrictive when it comes to the presented client certificates. This is done by inspected the presented certificates during the handshake and verifying that the expected CA certificate is among the certs presented.