E2E test validate peer certificate chain during TLS handshake (#4064)
Work around the issue that peer certificates are only populated in connection state after a successful handshake and that Elasticsearch 6.x is more restrictive when it comes to the presented client certificates. This is done by hooking into the TLS handshake and making sure the presented peer certificates can be validated against the expected CA certificate but do not require the full TLS handshake to succeed.