Support both kind 0.8 and 0.9
This is allow testing of all supported k8s versions. Given that we still want to support k8s 1.12 and 1.13, kind 0.8 is the only option to to test against those versions. At the same time we want to test the latest version of k8s as well even before they are available on the Cloud providers, for that we need kind 0.9