Improve upgrade test harness (#4060)
Keep one stack through all tests and upgrade it instead of checking the status of the old stack and then recreating a new one.
This is to address the edge case where a managed resource becomes unhealthy due to an operator upgrade but that situation is not detected because the stack is deleted too quickly.

By keeping the existing stack around we now have a single Elasticsearch cluster that lives through all upgrade tests. Any deterioration in health should be detected at the latest when the test harness is trying to upgrade the stack.

There is one exception for the alpha version of ECK. Resources created by the alpha version are not reconciled by later versions of the operator so we delete the alpha stack after upgrading to the beta.

A downside of this approach is that we are now testing two things together: operator upgrades and stack upgrades.