Add EKS ARM e2e job (#4000) (#4126)
Add an E2E job to run on ARM machines. This PR only introduces the job. After it is tested manually, it will be added to the nightly job in a separate PR. 

Introduces newer version of eksctl that should help with the test failures and ARM functionality is introduced in 1.4 so test job should be there as well.
Co-authored-by: Charith Ellawala <52399125+charith-elastic@users.noreply.github.com>