E2E: Extract Support Diagnostics from Elasticsearch clusters under test (#4317)
This adds another post-failure action to our e2e test setup to run the official support diagnostics tools against all currently deployed Elasticsearch clusters.

It is using a "naked" k8s Pod that runs the diagnostics in an init container and simply sleeps in the main container. This allows us to know when the diagnostics are ready (== when the Pod is ready). We then extract the results with kubectl cp and delete the Pod.