Make OCP deployer CI compatible (#4422)
The first iteration of this revised OCP driver was not really CI compatible:
We cannot use eck-dev images in CI => change to eck-ci.
We need to login into to docker because we cannot rely on a Docker login happening outside of the deployer invocation.
We need to use a shared volume between CI container and client container, which we don't need in local mode because the Make process is not running in a container as well.
We need to specify the OCP version in the Jenkins job we want to test. To conserve resources if the version is not specified it just picks one of the supported versions which are tested round robin during the course of a week.