Allow users to spec a serviceName in associations (#4404)
Add a new serviceName attribute to all associations (even though it makes only sense for Elasticsearch but the tradeoff here is additional complexity in the code). Query for the referenced service, extract the configured port which matches the currently configured protocol of the managed resource (e.g. Elasticsearch running with HTTPS) and use it to construct a service URL which will be configured in the associated resource.

Users need to create a service either ahead of time or once the Elastic stack resource have been deployed.
Users need to follow the port naming convention we use in ECK i.e. https or http (which is compatible with Istio's port naming convention btw)