Move kind bootstrap into deployer (#4491)
Kind is currently a bit of a snowflake in our e2e test setup. This aims at moving it under the deployer umbrella. The motivation is mainly to have a uniform way of installing manifests (with Kubernetes version awareness).

Drops the image loading functionality we had in the stand-alone version. I believe we don't need it anymore given that our images are all public and at least in my tests kind was able to pull them.