Support custom CA on HTTP layer  (#4575)
Allows users to add a full CA certificate in ca.crt and ca.key which will be used to sign the leaf certificate that ECK will use for the Elastic stack applications.

Awkwardly the options we offer in selfSignedCertificate can then still be used to modify the certificate. Awkwardly, because that certificate is no longer necessarily self-signed, depending on the CA the user has provided. But I don't think we want to break the stable API we have for that.

This also changes the naming convention for the transport layer CA user can specify for Elasticsearch to follow the new ca.crt, ca.key scheme. The old tls.key and tls.crt names are still supported for backwards compatibility.

Co-authored-by: David Kowalski <50632861+david-kow@users.noreply.github.com>
Co-authored-by: Michael Morello <michael.morello@gmail.com>