Elasticsearch monitoring with Metricbeat and Filebeat as sidecars (#4528)
Adds a new monitoring field to the Elasticsearch resource to configure one or two different Elasticsearch references to set up stack monitoring with Metricbeat and log delivery with Filebeat. The referenced ES are used to send the data collected by the beats.

```
apiVersion: elasticsearch.k8s.elastic.co/v1
kind: Elasticsearch
spec:
  monitoring:
    metrics:
      elasticsearchRefs: 
        - name: m1
          namespace: m1
    logs:
      elasticsearchRefs:
        - name: m2
          namespace: m2
```

This is implemented with a multiple association of ES type (1 es <-> [1|2] es).

New stackmon packages contains functions to:
- set the xpack.monitoring.* settings in the ES config
- enable stack logging for the ES container using the environment variable ES_LOG_STYLE=file
- add volumes to the ES pod to mount the ES CA and the beats configurations in the beats sidecar containers
- inject Metricsbeat and Filebeat as sidecar container in the ES pod than that of ES container with volumeMounts for the ES CA
- reconcile the Metricsbeat and Filebeat configurations in ConfigMaps

The beats configuration is built from a base configuration merged with the output config section which defines ES info to send data. For Metricsbeat, the base config is a template to inject the ES info to collect data.

A hash of the two beats config files is added in a pod label to ensure pods are rotated when es user passwords are rotated.

Limitations
- Minimum supported Stack version is 7.14.0 (to benefit from ES_LOG_STYLE=file)
- To use custom beat images, you have to override the podTemplate.
- The monitored Elasticsearch is not deployed until monitoring clusters are not ready yet or the association are not configured
- monitoring.[metrics|logs].elasticsearchRefs accepts only one Elasticsearch reference. It's a slice to future proof the API for Elastic agent.