Avoid circular dependency between Elasticsearch and association controllers (#4626)
The Elasticsearch controller waits until the assocation are ready to reconcile Elasticsearch, including creating the associated ES services and secrets. In parallel, the assocation controller waits for the ES service/secret to be configured.

This commit moves the association.AreConfiguredIfSet check present in the Elasticsearch controller after the services/secrets creation to let the ES controller create them and not block the association controller which can thus be configured.