Fix to allow to test EnterpriseSearch with Elasticsearch 8.0.0-SNAPSHOT (#4713)
Previsouly in #4604, I added WithEnvVar to the EnterpriseSearch builder to add the environment variable ALLOW_PREVIEW_ELASTICSEARCH_8X=true.
This adds the environment variable passed as a parameter in all containers defined in the pod template of the EnterpriseSearch resource. But if there is no containers in the pod template, we don't add the env var.

This commit fixes that by adding the main enterprise-search container if there is no container defined in the pod template.