Remove non-functional pre-stop hook parts (#4801)
The pre-stop hook script intended to improve availability during rolling upgrades has not been doing what it is advertising to do since ECK 1.3.0. 

We are including non-ready Pods in the headless service which we then try to use to determine when a terminating Pod has been removed from DNS (never, in this constellation). We have since added functionality in that relies on this setting for client-side node discovery (aka sniffing).

This means that our pre-stop hook was always running the full 20+30 secs before terminating the Pod. 
This change simplifies the pre-stop hook and turns it into a simple timeout without trying to check DNS for the Pod IP. I have opted for sticking with 50 seconds of wait time.  I also kept the terminationGracePeriod at 180s to avoid a rolling restart on ECK upgrade.