Improve Elasticsearch cluster health API call when timeout (#4803)
This commit updates the `request` function to accept an optional `skipErrFunc func(error) bool` argument to know if we should decode the response even when the status code is not 2xx.

If the Elasticsearch endpoint "/_cluster/health?wait_for_events=languid&timeout=0s" returns 408 (Timeout) then `GetClusterHealthWaitForAllEvents` swallowed the error and returned an empty/default value. 

We are now trying to decode the body as a best effort even when there is a 408 Timeout error for this specific ES API call.