Add support for Tanzu to deployer (#4846)
This is using a private container repository for the installer image therefore deviating from the concept of building client images on the fly as needed which we adopted for kind and ocp. This is mainly because we cannot share the Tanzu installer in a publicly accessible Docker image due to T&C restrictions around redistribution. Using this image we spin up a management cluster in Azure from which we can spin up the workload cluster which is then used to run our e2e test suite.