Shorten the reconciliation loop duration if Elasticsearch is down (#4938)
This commit allows for a reconciliation loop that takes 30sec instead of 1min30sec when Elasticsearch is down and immediatly update the Elasticsearch status instead of waiting another reconciliation loop.

The first call to the ES API is now to get the license (extracted from the license reconciliation) and is used to update the `esReachable` bool. Then we skip the steps that requires ES up (remote clusters settings, update annotation with cluster uuid) using `esReachable`. The Elasticsearch status is updated as soon as we got an error when calling ES with the last observed state.
`observedState` is changed to be able to get dynamically the last observer state just before updating the status, otherwise we have to wait for another reconciliation because at the beginning of the ES reconciliation that will fail because ES is down, the last observer state is very often still green because the state was retrieved between 0 and 10 secondes ago.