Avoid updating the association status when no association (#4986)
This fixes a bug where the association status of the Elasticsearch resource is updated by the EsMonitoring association controller when there is no association and therefore no changes to be made. The `oldStatus` and the `newStatus` appeared to be different because one was nil and the other was empty. The equality tests now supports this situation.