Support JSON marshalling of commonv1.Config on value receivers (#5056)
Updates MarshalJSON() method of commonv1.Config to use a variable receiver to make json.Marshal work with value and pointer receivers and adds unit tests.

Mixing pointer and non-pointer receivers is not a recommended pattern, mainly because it brings confusion for readers. Here it is only for one method which is not called directly, so it's ok.