Introduce orchestration hints as a way to gate orchestration features (#5072)
Introducing a new annotation to store orchestration hints. Use this field to make decisions about orchestration features that are relevant or not (currently enabling allocations and use of transient settings which are both not needed anymore when using node shutdown API). Not using a single hint value here to allow future extensions.
If lost will run the logic that added the hints again and recreate the annotation.