Change upgrade path validation for 8.0 to only allow 7.17 (#5261) (#5269)
This sets the minimal supported version for 8.0 to 7.17. It also changes the validation on version upgrade to take the .status.Version into account instead of the spec.Version (which it still uses as a fallback in case we do not have a valid .status.Version). The rationale behind that is that we want to make sure that users do update the version of the spec in quick succession without waiting for the rollout of the upgrade to complete thereby unwittingly circumventing the version restriction.