E2E: adjust continuous health check for non-HA clusters (#5279)
This adjusts the skip condition on the continuous health check we run during mutation tests to account for non-HA clusters. A single node or two master node configuration will loose availability due to loss of quorum during a rolling upgrade and we did not account for that in our existing checks.