Allow users to disable version validation to downgrade clusters  (#5272) (#5280)
Adds support for a new annotation called k8s.eck.elastic.co/disable-downgrade-validation as a troubleshooting tool to allow users that accidentally upgraded to the incorrect version to correct this mistake and go back to the previous version as long as Elasticsearch has not booted up and converted the data directory. This can be the case if the user upgraded to a version of Elasticsearch that does not exist or that refuses to boot because it is not on the upgrade path from the current version.