Only configure Stack Monitoring if association reconciled (#5339)
Currently, if a referenced monitoring Elasticsearch does not exist, Stack Monitoring is still configured with empty values in beat configs causing the beats to enter in state `CrashLoopBackOff`.
This happens only for ES-ES. For KB-ES, the call to `AllowVersion()` checks that `assoc.AssociationConf() != nil`, so the reconciliation is aborted if the association is not yet configured. This commits fixes both for consistency.