Do not upgrade all Elasticsearch nodes of a given tier at once (#5452)
Introduces a new upgrade predicate to prevent deletion of all the nodes of a given tier. Tier to be understood here as a generalisation of the existing "data tier" notion to all non-master roles.

Master nodes (and voting only nodes) are excluded as we are handling those already in existing predicates and generally apply much stricter rules to them (just one master node at a time can be upgraded no matter how permissive the change budget)

The focus here is on the existing nodes in the cluster and not on the expected nodes. My rationale is that the purpose of this predicate is to prevent degradation of the existing service for which the future state of the service is irrelevant.