Avoid DELETEs of non-existent secrets (#5488)
This change guards DELETE calls on secrets (cluster license, secure settings, legacy transport certs) with a GET call through a new shared utility function to remove unnecessary load on the API server.

It also renames keystore.NewResources to keystore.ReconcileResources. The reason for this rename was that while I was tracking down these DELETES, I was confused and surprised to find reconciliation activity happening in a function that appeared to be a simple constructor with New* in the name.