Reset phase on each reconciliation in Elasticsearch status (#5507)
Instead of carrying the previous phase over from persisted state reset the phase on each reconciliation to a neutral value (""). Because we set a phase value even in the presence of errors this should ensure that phase values like Invalid that we give priority over other values do not stick around forever. Also move the priority logic into the state struct method.