Support custom secret for associations with external Elastic resource (#5240)
This adds support for using a custom secret to associate an ECK-managed Elastic resource with an external non-ECK-managed Elastic resource.

Changes:
- Add a field secretName: string to the common.ObjectSelector struct used to reference a resource in an association. This change impacts the API of all the Elastic resources.
```
  elasticsearchRef:
    secretName: test-ext-es-ref
```
- Use an ObjectSelector instead of a NamedNamespace for the type of the key of the map of AssociationConf (used in assocations that support multiple references) to take in account the new field.
- Change the signature of ReferencedResourceVersion: func(c k8s.Client, esRef commonv1.ObjectSelector) and implement for all types of Elastic resources how to get the version of the referenced external resource (requesting its HTTP API and extracting the version from the response using a jsonPath).

Basically the main change consists to shortcut the reconciliation loop of the association controller, if a referenced object is a of type Secret. Then, get the version of the referenced external resource, which is the opportunity to check that we can reach it and create an association config (which stored in an annotation of the associated resource) with the auth and CA Secrets pointing to the custom Secret..

    "association.k8s.elastic.co/es-conf-620140190": "{
	\"authSecretName\":\"test-ext-es-ref\",\"authSecretKey\":\"password\",
	\"caCertProvided\":false,\"caSecretName\":\"\",
	\"url\":\"https://test.es.abc-0.xyz.com:1234\",
	\"version\":\"7.16.3\"}",

In the Stack Monitoring association 'es-es', we configured MetricBeat with a predefined user that exists when the referenced ES is managed by ECK. This is changed to use the username and password defined in the custom secret instead.

Before FleetServer 8.0, there is no API to get the version, so we just set unknown for the version and the compatilibity check is skipped.

The user corresponding to the username, password referenced in the custom secret must have the appropriate privileges depending on the type of association. When the referenced resource is managed by ECK, a user with a specific role with the right privileges is created specifically for the association.