Annotate operator pod to speed up webhook certificate secret propagation (#5519)
When the operator is deployed for the first time, the Secret which contains the webhook certificate is populated
 by the operator (if the webhook certificate is managed by ECK, which is the case by default). Unfortunately it can
take some time for the content of the Secret to be propagated into the operator container. This commit marks the
operator pod as updated using an annotation to speed up the secret propagation.