Support a globally shared CA (#5539)
This adds support for a new operator parameter that allows users to specify a shared CA that will be used for all managed resources. If users opted to specify separate CAs per resource those configurations will be left untouched, however any resources that are using the default per-resource self-signed CA will use the shared CA.

The CA is specified as a directory via an operator param, --ca-dir. The operator expects two files to be present in this directory tls.crt and tls.key -- ca.crt and ca.key are supported as well to be consistent with the existing per-resource custom CA feature.

The operator watches the directory and restarts if it detects changes.