Do not delete restart shutdowns before node rejoins cluster (#5618) (#5623)
Predicates the deletion of a node shutdown record of type RESTART on the presence of the corresponding node in the cluster unless cluster is fully reconciled. Presence in the cluster is established via the cached node list maintained in the operator and refreshed on each reconciliation.

Adds flexible predicates to clear. This way we can call it multiple times: once we have observed all expected deletions to clean up all shutdowns for the happy path where the node has rejoined the cluster. We then call it a second time behind all the conditional logic that we also apply to the "should we re-enable shard allocations" question in the pre-shutdown world. This second time is only predicated on completed shutdowns so once we reach a stable state we also delete shutdowns for nodes that did not rejoin the cluster. This would be the case for example if a node lost its state during restart du to an external event e.g. a user manipulating the node's data directory or a hardware failure. It is debatable whether this optimisation is necessary as we could simply only delete shutdown once we reach a stable state (which is btw what happens during scale downs). The only downside would be that shutdown records start piling up in the Elasticsearch cluster state. I could potentially be a problem on very large clusters with hundreds of nodes. The optimisation is more or less free for us so I went for it here.

Co-authored-by: Michael Montgomery <mmontg1@gmail.com>
Co-authored-by: Michael Morello <michael.morello@gmail.com>
(cherry picked from commit ca1d5b597bf00cdaf6a1d13e546de25be2e8ff0e)