E2E: retry check cluster UUID check (#5646)
The "Cluster UUID should have been preserved" check sometimes fails if the connection to Elasticsearch times out. While this might point to an underlying issue with the Elasticsearch cluster, this step is not intended to test Elasticsearch availability. This commit thus wraps the step in test.Eventually to allow retries.