Allow providing cleartext passwords for creating ES users (#5613)
Add support for an additional form of secret following the https://kubernetes.io/docs/concepts/configuration/secret/#basic-authentication-secret format. It will allow users to specify username and a cleartext password.

I also added support for a custom roles file in that secret. The idea being that a user that wants to create a user and assign a role to it does not need to create two separate secrets for that (one for the user and one for the users_role file). Instead they can list the roles to be assigned to the new user in the same secret as a comma separated list.

We always supported setting a custom password for the elastic user through the file realm (because user defined ES users always take precedence) This change also removes the default elastic user secret if the elastic user is set up through the file realm. The rationale for this decision was to avoid having to sync the contents of both secrets (the user provided one and the operator created one) or have the contents diverge which might lead to confusion. Also we should not not create resources in the API server that are not needed

Co-authored-by: Thibault Richard <thbkrkr@users.noreply.github.com>
Co-authored-by: Michael Morello <michael.morello@gmail.com>