E2E: Skip node shutdown check for non-HA clusters (#5671)
When I added the node shutdown check to the e2e tests in #5642 I ran all tests via our PR e2e job but this runs them on the current stack version i.e. 8.2.0 so all version upgrade tests were not run. This hid an incorrect assumption I had made about the legitimate number of restarts that can exist at a given time. When we upgrade non-HA clusters we restart all nodes at once so it is OK to have more restarts in flight at the same time than the change budget allows.

I opted to disable the check as a whole on small clusters instead of adjusting the formula to keep the calculation simple.

Co-authored-by: Thibault Richard <thbkrkr@users.noreply.github.com>