Support disable-downgrade-validation for all relevant apps (#5680)
Support the annotation that disables the downgrade protection on all relevant apps. That is all apps other than Elastic Maps Server where the downgrade restriction does not apply.

Also added a sentence to troubleshooting guide. This is on purpose in the "I upgrade Elasticsearch by mistake" section because in all other cases a forced downgrade is not possible. Only if our association mechanism has effectively prevented the upgrade from actually being rolled out (because we are still waiting on Elasticsearch) is a downgrade possible. If let's say Kibana had been accidentally upgraded when it should not have been then there is no way back as the corresponding indices will have been migrated by the new version of Kibana already. So this PR closes effectively only a tiny gap in the downgrade workaround that was not covered yet.