Allow non-IPs in service spec to avoid noop updates (#5663)
We are copying parts of the service spec populated on the service side into the expected service spec to avoid unnecessary updates of services. For some reason that is unfortunately not documented an exception was made for the ClusterIP[s] values that not valid IPs like None. By not including those we always update the service on each reconciliation. This change addresses this gap and a similar one for internalTrafficPolicy

Co-authored-by: Michael Morello <michael.morello@gmail.com>
Co-authored-by: Thibault Richard <thbkrkr@users.noreply.github.com>