License check annotation  (#5815)
Support an annotation on all Elastic resources to request a certain license level in the operator. Pre-req of #5781

I had to move a few things around for this to break a few import cycles, I hope that the changes and new packages names are self-explanatory and generally an improvement over the status quo.