Add note for air-gapped environments for pulling the Support Diagnostics docker image
A lot of ECK clusters are air-gapped, meaning they have no access to our docker.elastic.co registry, so we should add a note as it's not immediately clear a manual actions is required for people that would need to get the Elasticsearch & Kibana diagnostics. I made a quick draft, feel free to edit the wording/phrasing.