Fix .env permission in CI pipelines (#6163)
Creating the .env file inside the ci docker container just to pass shellcheck was a bad idea. The file is created as root and this prevents setenvconfig which is run outside the docker ci container from updating it.
This commit removes the creation of .env in the shellcheck make target and instead moves the setenvconfig call before make ci-check.