New CRD StackConfigPolicy to declaratively configure multiple Elasticsearch clusters (#6148)
New custom resource StackConfigPolicy to configure cluster settings, snapshot repositorires and lifecycle policies for a list of Elasticsearch clusters matching a labels selector.

```
apiVersion: stackconfigpolicy.k8s.elastic.co/v1alpha1
kind: StackConfigPolicy
metadata:
  name: config-staging
  namespace: elastic-system
spec:
  resourceSelector:
    matchLabels:
      env: staging
  elasticsearch:
    clusterSettings:
      indices.recovery.max_bytes_per_sec: "100mb"
    snapshotRepositories:
      backup:
        type: gcs
        settings:
          bucket: "gcs-bucket"
    snapshotLifecyclePolicies:
      test-snapshots:
        schedule: "0 1 2 3 4 ?"
        name: "<staging-snap-{now/d}>"
        repository: "backup"
        config:
          indices: ["*"]  
          ignore_unavailable: true
          include_global_state: false
        retention:
          expire_after: "30d"
          min_count: 1
          max_count: 50
```

The namespaces used to find Elasticsearch clusters to configure depend on the StackConfigPolicy namespace. If it is the operator namespace, all namespaces managed by the operator are used, otherwise only the StackConfigPolicy namespace.

A new FileSettings Secret `<esName>-es-file-settings` owns by Elasticsearch is created empty and mounted as data volume by the Elasticsearch controller. The StackConfigPolicy controller soft owns the Secret and only updates it. On deletion, soft owned labels are used to find Secrets to reset.

For the SecureSettings Secrets defined in the StackConfigPolicy, the StackConfigPolicy controller writes the Secret namespaces and names in an annotation of the FileSettings Secret. The ES controller watches and reads the Secrets from that and merges their content with the user-provided Secrets of the Elasticsearch resource in the existing Secret `<esName>-es-secure-settings`.