Fix version gate for the file-based settings volume (#6289) (#6293)
This commit fixes an issue when upgrading to 8.6.0. We should not use the lowest version in the pods and instead the version of the Elasticsearch resource to reconcile the empty File based settings Secret if it doesn't exist, otherwise the upgrade is blocked because the new pods are created with a new volume waiting for the Secret to be present.