Refactoring vault package used by tooling (#6418)
This adds a new vault package to load secrets for the license generator used in go generate, the Helm releaser and the e2e-tests runner.

The current behaviour useful in dev is not changed: if the file exists, nothing is done, if the file doesn't exist, it is retrieved from Vault and written on disk.
A new mode is added which fetches the file from Vault and keeps it in memory which we will use in CI.

- TEST_LICENSE="" // no secret read
- TEST_LICENSE=$(pwd).ci/test-license.json // secret read and written to the given path
- TEST_LICENSE=in-memory // secret read but not written to disk

It also simplifies configuration of secrets which require different path depending on context (k8s agent or docker in gcp agent) because when we are in docker we set the path before we are in docker. With that, no path to define, we just read in memory.