Fix handling of unmanaged esRef in Beats stack monitoring (#6482)
Beats stack monitoring assumed that the associated Elasticsearch cluster is a local cluster managed by ECK. However a Beats manifest with external references would break the stack monitoring feature. Worse still we would silently swallow the error creating the appearance of normalcy when the situation was in fact not normal. 

Address these issues by:
1. handling the error correctly 2. special casing external references to Elasticsearch correctly by
distinguishing between external references and internal references when generating Beats configuration making a request in case of a remote unmanaged cluster to find out the cluster UUID which we seem to need for stack monitoring configuration (I have not tested what happens if we don't set it ). 

Known issues:
If the external Elasticsearch cluster has stack monitoring for the cluster itself turned off then the Beats monitoring data will not show up in the Kibana UI (a known issue that we already document). However monitoring data will still flow into the cluster for use in custom dashboards etc.