Move setenv in snapshot-versions Jenkins job (#6517)
This fixes a side effect of #6465 which causes an issue in the Jenkins job
cloud-on-k8s-e2e-tests-snapshot-versions by ensuring the environment 
is set before any call to make -C .ci ci.

The first step make -C .ci TARGET=validate-jenkins-pipelines ci causes the
fetch of the prod public license key.
Next .ci/setenvconfig dev/build sets BUILD_LICENSE_PUBKEY=dev. Because
the .ci/license.key already exists, the dev license public key to test snapshots
is not retrieved.

This is more of a quick fix as this job is revisited on Buildkite.