Buildkite release pipeline (#6377)
This adds the missing part to the buildkite pipeline to be able to release ECK.
First, it adds build steps to:
- build multi-arch images (amd/arm)
- build ECK in fips mode
- also build ECK with a dev license pub key (not necessary for the release but needed to test snapshot version of the stack)
I moved all these steps in a dedicated pipeline-build.yml.

Until this PR, the images were built using buildah. Building multi-architecture containers with Buildah is possible but is not yet possible on Buildkite. That's why I switched back to a gcp agent to continue to use docker buildx.

I put the set of steps for the release in a new pipeline-release.yml that does the 3 parts: k8s manifests upload, release the helm charts and release eck to the redhat operatorhub.

I rewrote the k8s manifests upload in a bash script to have one simple script that fetches the secret and upload the files that we can run using a bk k8s agent instead of relying on the .ci/Makefile, the ci container and the hack/manifest-upload.sh script.