Revisit assertion to not rebuild image
If the image is built for one platform (in the context of a PR or main)
then we cannot use `grep -q 'linux/arm64'` to verify that the image has
already been built. We can't even use `grep -q 'linux/arm64'` because
if there is only one platform, platforms are not listed.