[ci] Release Helm charts after operator build (#6594)
Currently the Helm charts are published before the operator images are published because it's faster. This is not a problem at the moment, but I think it's better and logical to publish Helm charts after the operator images. It avoids the very rare case where you update the helm repository during the release and end up with an operator pod that won't start because the image hasn't been published yet.